requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
alabaster==0.7.13
    # via sphinx
asttokens==2.2.1
    # via stack-data
babel==2.12.1
    # via sphinx
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
    # via furo
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
coverage==7.3.0
    # via -r requirements.in
decorator==5.1.1
    # via ipython
docutils==0.20.1
    # via sphinx
executing==1.2.0
    # via stack-data
furo==2023.8.17
    # via -r requirements.in
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
ipython==8.14.0
    # via -r requirements.in
jedi==0.19.0
    # via ipython
jinja2==3.1.2
    # via sphinx
markupsafe==2.1.3
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
packaging==23.1
    # via
    #   pytest
    #   sphinx
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==1.3.0
    # via pytest
prompt-toolkit==3.0.39
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyclean==2.7.4
    # via -r requirements.in
pygments==2.16.1
    # via
    #   furo
    #   ipython
    #   sphinx
pytest==7.4.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   sphinx
six==1.16.0
    # via asttokens
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==7.2.5
    # via
    #   -r requirements.in
    #   furo
    #   sphinx-basic-ng
    #   sphinxcontrib-applehelp
    #   sphinxcontrib-devhelp
    #   sphinxcontrib-htmlhelp
    #   sphinxcontrib-qthelp
    #   sphinxcontrib-serializinghtml
sphinx-basic-ng==1.0.0b2
    # via furo
sphinxcontrib-applehelp==1.0.7
    # via sphinx
sphinxcontrib-devhelp==1.0.5
    # via sphinx
sphinxcontrib-htmlhelp==2.0.4
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.6
    # via sphinx
sphinxcontrib-serializinghtml==1.1.9
    # via sphinx
stack-data==0.6.2
    # via ipython
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
urllib3==2.0.4
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
